Move all routes to api/route.ts
